+++ app/controllers/todo_categories_controller.rb
@@ -12,7 +12,11 @@ class TodoCategoriesController < ApplicationController
 
   # GET: /todo_categories/5
   get '/todo_categories/:id' do
-    TodoCategory.find_by_id(params[:id]).to_json
+    if TodoCategory.find_by_id(params[:id])
+      TodoCategory.find_by_id(params[:id]).to_json
+    else
+      {error: "Record not found with id #{params['id']}"}.to_json
+    end
   end
 
   # PATCH: /todo_categories/5
@@ -27,7 +31,11 @@ class TodoCategoriesController < ApplicationController
   # DELETE: /todo_categories/5/delete
   delete '/todo_categories/:id' do
     todo_category = TodoCategory.find_by_id(params[:id])
-    todo_category.destroy
-    todo_category.to_json
+    if todo_category
+      todo_category.destroy
+      {message: "Record successfully destroyed"}.to_json
+    else
+      {error: "Record not found with id #{params['id']}"}.to_json
+    end
   end
 end
